requirements.txt
asgiref==3.8.1
Django==5.2.1
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-role-permissions==3.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
